# api/storage/api-docs/api-docs.yaml
openapi: 3.0.0
info:
  title: 'Pixel Api'
  version: 1.0.0
paths:
...
  '/api/cabinets/{cabinetId}/domains/{id}':
    get:
      summary: 'Конкретный домен'
      operationId: 'App\Http\Controllers\DomainController::show'
      parameters:
        -
          name: cabinetId
          in: path
          required: true
          schema:
            type: integer
        -
          name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        200:
          description: 'successful operation'
    put:
      summary: 'Обновить домен'
      operationId: 'App\Http\Controllers\DomainController::update'
      parameters:
        -
          name: cabinetId
          in: path
          required: true
          schema:
            type: integer
        -
          name: id
          in: path
          required: true
          schema:
            type: integer
        -
          name: name
          in: query
          required: true
          schema:
            type: string
        -
          name: domain
          in: query
          required: true
          schema:
            type: string
        -
          name: 'on'
          in: query
          required: true
          schema:
            type: bool
        -
          name: token
          in: query
          required: true
          schema:
            type: string
      responses:
        200:
          description: 'successful operation'
        403:
          description: Forbidden
    delete:
      summary: 'Удалить домен'
      operationId: 'App\Http\Controllers\DomainController::destroy'
      parameters:
        -
          name: cabinetId
          in: path
          required: true
          schema:
            type: integer
        -
          name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        200:
          description: 'successful operation'
        403:
          description: Forbidden
  '/api/cabinets/{cabinetId}/members/me':
    get:
      summary: 'Текущий сотрудник'
      operationId: 'App\Http\Controllers\MemberController::currentMember'
      parameters:
        -
          name: cabinetId
          in: path
          required: true
          schema:
            type: integer
      responses:
        200:
          description: 'successful operation'
...
  '/api/cabinets/{cabinetId}/members/{id}':
    get:
      summary: 'Посмотреть сотрудника'
      operationId: 'App\Http\Controllers\MemberController::show'
      parameters:
        -
          name: cabinetId
          in: path
          required: true
          schema:
            type: integer
        -
          name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        200:
          description: 'successful operation'
    put:
      summary: 'Обновить сотрудника'
      operationId: 'App\Http\Controllers\MemberController::update'
      parameters:
        -
          name: cabinetId
          in: path
          required: true
          schema:
            type: integer
        -
          name: id
          in: path
          required: true
          schema:
            type: integer
        -
          name: roleId
          in: query
          required: true
          schema:
            type: integer
      responses:
        200:
          description: 'successful operation'
        403:
          description: Forbidden
    delete:
      summary: 'Удалить сотрудника'
      operationId: 'App\Http\Controllers\MemberController::destroy'
      parameters:
        -
          name: cabinetId
          in: path
          required: true
          schema:
            type: integer
        -
          name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        200:
          description: 'successful operation'
        403:
          description: Forbidden
...
  /api/users/me/change-password:
    post:
      summary: 'Список кабинетов'
      operationId: 'App\Http\Controllers\UserController::changePassword'
      parameters:
        -
          name: passwordOld
          in: query
          required: true
          schema:
            type: string
        -
          name: passwordNew
          in: query
          required: true
          schema:
            type: string
      responses:
        200:
          description: 'successful operation'
        500:
          description: 'Invalid creditionals'
security:
  - []